+++ src/main.rs
@@ -88,6 +88,7 @@ fn mainloop() -> Result<(), Box<dyn Error>> {
                             config.servers_pull,
                             config.mirror_pull_interval.as_secs(),
                             config.chunk_size_changes,
+                            config.chunk_size_checks,
                         );
                         cleanup::start(
                             config.connection_string.clone(),